Move robotoff.insights.ocr to robotoff.prediction.ocr
