+++ robotoff/prediction/ocr/brand.py
