+++ robotoff/prediction/ocr/core.py
