+++ robotoff/prediction/ocr/dataclass.py
