+++ robotoff/prediction/ocr/image_flag.py
