+++ robotoff/prediction/ocr/image_orientation.py
