+++ robotoff/prediction/ocr/location.py
