+++ robotoff/prediction/ocr/product_weight.py
