+++ robotoff/prediction/ocr/store.py
