+++ robotoff/prediction/ocr/utils.py
